Use a prelude to define the base types in notlang.
We create a prelude and put some basic types in, without
any implementation as that remains in the interpreter.

There are still some things assumed by the type checker,
such as the types of the binary operations.

These changes led to reworking the way that assignments
are handled, where we track the type of a variable as it
varies rather than having one TypeExpr that is the type of that
variable across its lifetime. This is incomplete as it
doesn't deal with branches currently.